norm almost done

+++ src/split.c
@@ -1,8 +1,20 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   ft_split.c                                         :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: maburnet <marvin@42.fr>                    +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2023/09/16 20:47:18 by maburnet          #+#    #+#             */
+/*   Updated: 2023/09/16 20:47:32 by maburnet         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #include "../include/push_swap.h"
 
 void	ft_bzero(void *s, size_t n)
 {
-	size_t	i;
+	int	i;
 
 	i = 0;
 	while (i < n)
@@ -14,8 +26,8 @@ void	ft_bzero(void *s, size_t n)
 
 void	*ft_calloc(size_t nmemb, size_t size)
 {
-	size_t		mem;
-	void		*ptr;
+	int		mem;
+	void	*ptr;
 
 	mem = size * nmemb;
 	if (size == 0 || nmemb == 0)
